Change extraction to match balancing
Here we match the balancing matching which happens with a wide
dataframe. So we take all the covariates and make sure that they are
saved as columns rather than in the long format.

This is happening in the notebook (analysis/extract_to_feather.ipynb),
but it should be eventually passed to main.

A second change is the balancing notebook with some grid-search done for
the CBPS