Solve W design matrix and drop treat columns
In the run_balancing.py function we change the design of the treatment.
In the past we were mislabeling treatments and classifying as treatment
those fires where the intensity was larger than the actual treatment.
Thus we do a equality restriction where the pixels treated multiplied by
the treatment must be identical to the intesity class of interest!

Also, remove columns from class_dnbr and class_frp that can be in the
dataframe and go to the balancing dataset.