Organize the house a bit and consolidate utils
This commit creates some modules to clean the house a bit and to keep
things separated and cleaner. Now, we still need to add a builder module
that takes care of the dataset creation rather than having things in
Jupyter notebooks. Classes can be messy, but also functions are annoying
as we need like 10 functions to build things.

Classes might be heavier?

TODO:
 - Add estimate module with all the CBPS code