Revamp processing so we can generate feather files
The processing skips the extra steps and process feather files from the
processing functions rather than rely on a notebook. This is easier if
we need to re-run stuff since notebooks are just annoying to run.

Ideally we can add some hydra configs to this, but are we running this a
lot? idk