Change balancing to use SQL and packaging
This commit trasnforms the balancing code so it uses a SQLlite database
rather than saving a bunch of files. The code will create model_id using
the lr and reg values to create a hash.

Additinally, a packaging file is created to make more experiments with
the running. Somethings are still waiting to be working as we change the
src as prescribed.

The weighting file is also moved to main.