Revamp package structure to make it work as a pkg
Change src to be src/prescribed. The idea is to allow pyproject to work
well during the installation process and use the prescrbed namespace
rather than just doing src. I am following the Python new packaging
tutorial, and not doing the classic setup.py and setup.cfg.

Now, some small changes are doing on the code to make things work ok and
a script is added to deal with the building of a dataset.