Add new main scripts for balancing
Ok, testing is on baby! Build and extraction are now *chef's kiss*. Now,
we need to balance stuff and a new main script to prepare data and do
the actual balancing is added here with its corresponding configuration
file. The idea is to test this and see if the SQLite database is working
or not!