Add db creation to pipeline
To have a more robust approach to SQLite, we create tables with primary
keys to avoid data repetition in case we need to re-run code or
something else. This is pretty nice as we might need to run models
several times depending on what we find.

Other changes include some aesthetics and problems with f-strings in the
balancing code.