Change database to store more data
So, we need to keep track of the land type that we use. The dabase
wasn't storing that data so this commit adds that column to the db
schemas and to the data creation itself. The way we do this is by adding
a generic dictorionary on the function, this is mean for land type, but
can be use for other kind of metadata with the caveat that SQL schemas
should be changed!

More changes:
 - Fix db creation typos.
 - Change main to add land type to dataset