Abandon SQLite :tired_face:
The whole things with WAL didn't work, SQLite is still blocking the
files and is honestly annoying! So concurrent writting is hard and
hostely not worth the time. Now, a better folder system is created to
store results and the hope is to use something like duckDB that can deal
with Parquet files efficiently to read results without the extra
overhead of the SQLite stuff. Sad.