Add emissions extraction function
We implement the extraction of emissions from Xu, et.al. (2022) paper.
This extraction is very similar to the dNBR extraction, so we just
modify this process a bit to make it similar. We add also some changes
to the config extraction to map to the main script.

A second big change is the addition of the estimator calculation, which
is just the ATT. This is the last step after calculating weights and do
the balancing

Also, we add some small changes:
 - Change of logic in FRP extraction