Add main script to select and save best models
This commit adds a simple script to read all model files in parquet
format using DuckDB and saves the best models using loss and the
absolute estandarized mean differences for each iteration of our models.
These outcomes can be used to create figures and run the estimations
given outputs as we also save the selected weights.