Extract NVDI from rasters
This is very much WIP. We have still some issues on how we process these
files. The resulting mosaics after the reprojection to 1 km from 30 m
have lines because the reprojection is not always creating a nice affine
to work with. We can change the order of operations, but that can have a
larger memory footprint.