Abandon multiprocessing :(
Sadly rasterio is not playing nice with doing all this in parallel.
Thus, all of the things are back to normal non-parallel processing. My
guess is that opening the same file is complicated and doing memory
sharing is too annoying to set up, so we go with quick and dirty.

To-do:
 - One day figure out parallel processing, rather than threads from GDAL
   on this baby.