Add scaling to estimates if needed
This commit changes the estimation function to use a scale if needed.
This is useful when we want to change estimates on the fly rather than
changing the data. An example of this is emissions that come in Kgs, but
we sometimes just care about metric tons, thus a 0.001 scale is needed
to have the estimates in the right unit. Not essential, just
consistency.