Pooling ATTs using OLS w/ JK
This commit adds a pooling strategy for the ATTs across lags. These new
functions fit a OLS function for all the ATTs and calculate the errors
of these relationship using a Jackknife approach. The plotting of these
results is modified to take the pooled results too, so we can just see
how these compare with just plotting the raw results :)

Some things might be escping me, so errors might be still there