Add a distance definition to define treatments
Using population and distance, we now can define treatments as fire
spillovers. This is using two different variables, one is population
along a buffer around a particular fire, and the second one is an
additional buffer to define how "long" are the spillovers. With this we
can redefine treatments for balancing.