FRP overhaul! Re-processing FRP and change mains
Well, we change how we extract FRP and make sure there is no weird stuff
in our data. Now the extraction is more direct and is not writing to
NetCDFs, but just keeping things simple and extracting directly.
Additionally, there is a new change in how we do confidence by just
keeping by default fire obs that are nominally confident!

This involves several changes too!
 - Now we can do two types of treatments: MTBS or FRP ones, which we
   divide in two different config files. The configuration can be run in
   Hydra by using the `--config-name=` option.
 - We also made some small changes in how we build the dataset, removing
   a lot of the staggered options, which are not relevant in this
   setting.
 - Update to notebooks to show data and also some changes in the `plot`
   function to keep severity and intensity thresholding constant across
   them  (1 always to signify low).