Modify configuration files and add more comments
This is a simple commit! We are using the full force of Hydra to do
configurations for all kind of treatments. This is still WIP and not
something that is particularly essential, but still worth looking.

Some comments are added and also some desired behaviors when it comes to
read in files when a string is passed rather than actual object