Change estimation w/ Poisson
This is a big change and is a comparable implementation as the one
suggested by Wu. We played around a lot and finally found a correct way
to do this estimation using the R code provided by them. We change the
estimation and now it returns the correct RRs and comparable results to
the balancing shown in the paper.

 - Interestingly enough, the relationship between Poisson and
   Quasipoisson is merely just the name as the changes happen in the
   variance estimation and that's not very relevant for us as we do the
   JackKnife.