Update conf files configuration
I want to exploit the full modularity of the Hydra configuration to
create better configs thata are more flexible and allow to have multiple
configurations for each experiment. Ideally the optim stuff is never
changed and we have different experiments for frp, dnbr experiments,
etc.

We just changed tyra and build. In the future we do the balancing and
hopefully a main main file to run all in one go.