Change high-intensity definition and fix kwargs
This commit just changes the high-intensity/severity definition in the
calculation of the RR esimtates. Rather than taking a "higher than", we
use a `.isin()` to define the levels with more precision. Also, we
change the create distances function to have more direct definitions of
the kwargs. More work needs to be done on this one to reflect treatments
in spillovers.