Create distance treatments: avoid missing data
This commit has a final version for the distance assignment treatments
for the spillover estimation. In a previous version of this function we
had a problem where only the rows with distance were preserved and when
trying to filtering, we were creating new columns which was problematic.
This code is baiscally changing that and keeping all the rows from
the template so all the merges are "safe".