Fix spillovers to get ATT and not only RR
In this commit we change the estimator calculation to give us the ATT
besides the RR estimation. This is useful to inform the simulations that
are dnbr based and where a probability change does not informs very well
things. We also update the notebook to reflect changes in dnbr and not
only on FRP. The RR results are the same.