Add spillovers to simulations
This commit adds a simulation option to the sampling process within the
simulation code. On one hand, we use the changes in estimation of the
ATTs for the spillovers to introduce them into the simulation. On the
other hand, we change the configuration so each Rx fire has an effect
while avoiding double counting for the effects.

TO-DOs
-----
 - Maybe some functions are still with incomplete docstrings
 - Create a main script to run simulations to not run code in a
   notebook.