Change smoke linking confidence_frp
Estimate the relationship between smoke and severity using Jeff's
matched data. We watimate the polynomial fit of the data using a split
train-test sampling approach and select the best fit for the data. This
function is going to inform the policy emissions response using the
simulated severity.

We want some plots from here: mainly the fit and the overlap in the
severity distribution of data