Changes in modeling and predictions for sims
Change simulation code to take into account parameter uncertainty from
pollution predictions. Now we sample from a normal distribution: N(est,
se) for each of the interations so we get a more accurate idea of the
effect of the treatment on emissions. We do get a huge uncertainty on
small treatments, so we need to check that.