Add hydra config for simulation runs
This commit modified our main script to run simulations reading from a
simple hydra configuration. The hydra configs are added to this commit
to. We want to keep this flexible as the simulations have a lot of
moving pieces that are not to be hardcoded in a file and argparse/click
would be equally annoying. We set some defaults to this function and as
the other configs, we rely on defaults for basic paths that do not
change across runs.