Revamp configs to map compute better
Configuration files had a lot of mess. Hopefully not anymore. This
separates compute so we can just change options in the hydra CLI
execution of the scripts. No need to have separate stuff per compute.
The user just needs to define the data paths in the conf/compute and
maybe maybe just change data/defaults.yaml if needed.