Update configs and small changes to spillovers
This commit changes configurations a bit to make them clearer. It also
fixes an error where Hydra was unable to do path interpolation because
of a missunderstanding on how we use variables in YAML. This tests and
builds the datasets and also does small changes to make the code
clearer.